♻️ refactor(mcp): align world-model.ts → world_model.ts (one-stem) + document onboard tools-only (#68)
