+++ mcp/trajectory-server/src/tools/README.md
@@ -23,7 +23,7 @@ The MCP tool modules for the trajectory server — one TypeScript file per domai
 | Config | `config.ts` | `config_get`, `config_set`, `config_list` |
 | Onboard | `onboard.ts` | `onboard_state_get`, `onboard_get_questions`, `onboard_apply` (+ `origin` helper) |
 | Scan | `scan.ts` | `scan_run` (forks `scripts/scan.sh`), `repos_list` |
-| World model | `world-model.ts` | `world_model_get`, `world_model_search` (kuzu graph) |
+| World model | `world_model.ts` | `world_model_get`, `world_model_search` (kuzu graph) |
 
 `onboard-hooks-shim.ts` is a non-tool helper: it writes the plugin's PreToolUse hooks into the user's `settings.json` so they fire in headless `claude -p`, where marketplace plugin hooks are absent.
