+++ mcp/trajectory-server/src/tools/index.ts
@@ -19,7 +19,7 @@ import { compositeTools } from './composites.js';
 import { onboardTools } from './onboard.js';
 import { scanTools } from './scan.js';
 import { cheatcodeTools } from './cheatcode.js';
-import { worldModelTools } from './world-model.js';
+import { worldModelTools } from './world_model.js';
 export let toolDefinitions: Tool[] = [];
 export let toolHandlers: Record<string, (args: Record<string, unknown>) => Promise<CallToolResult>> = {};
