+++ mcp/trajectory-server/src/tools/world_model.ts
